+++ Demo/ApplozicSwiftDemoUITests/ApplozicSwiftGroupSendMessageUITest.swift
@@ -20,7 +20,7 @@ class ApplozicSwiftGroupSendMessageUITest: XCTestCase {
     override func setUp() {
         super.setUp()
         continueAfterFailure = false
-        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessNotificationInApplication) { (alerts) -> Bool in
+        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessNotificationInApplication) { alerts -> Bool in
             if alerts.buttons[AppPermission.AlertButton.allow].exists {
                 alerts.buttons[AppPermission.AlertButton.allow].tap()
             }
@@ -51,7 +51,7 @@ class ApplozicSwiftGroupSendMessageUITest: XCTestCase {
         let openPhotos = app.buttons[InAppButton.ConversationScreen.openPhotos]
         waitFor(object: openPhotos) { $0.exists }
         app.buttons[InAppButton.ConversationScreen.openPhotos].tap() // Click on photo button
-        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessPhoto) { (alerts) -> Bool in
+        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessPhoto) { alerts -> Bool in
             if alerts.buttons[AppPermission.AlertButton.allowAllPhotos].exists {
                 alerts.buttons[AppPermission.AlertButton.allowAllPhotos].tap()
                 return true
@@ -77,7 +77,7 @@ class ApplozicSwiftGroupSendMessageUITest: XCTestCase {
         let openContact = app.buttons[InAppButton.ConversationScreen.openContact]
         waitFor(object: openContact) { $0.exists }
         openContact.tap() // Click on Contact button
-        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessContact) { (alerts) -> Bool in
+        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessContact) { alerts -> Bool in
             if alerts.buttons[AppPermission.AlertButton.ok].exists {
                 alerts.buttons[AppPermission.AlertButton.ok].tap()
                 return true
@@ -96,7 +96,7 @@ class ApplozicSwiftGroupSendMessageUITest: XCTestCase {
         let openLocation = app.buttons[InAppButton.ConversationScreen.openLocation]
         waitFor(object: openLocation) { $0.exists }
         openLocation.tap() // click on location button
-        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessLocation) { (alerts) -> Bool in
+        addUIInterruptionMonitor(withDescription: AppPermission.AlertMessage.accessLocation) { alerts -> Bool in
             if alerts.buttons[AppPermission.AlertButton.allowLoation].exists {
                 alerts.buttons[AppPermission.AlertButton.allowLoation].tap()
                 return true