+++ Sources/Views/ALKFriendVideoCell.swift
@@ -49,6 +49,7 @@ class ALKFriendVideoCell: ALKVideoCell {
 
     override func setupViews() {
         super.setupViews()
+        let width = UIScreen.main.bounds.width
 
         let tapGesture = UITapGestureRecognizer(target: self, action: #selector(avatarTappedAction))
         avatarImageView.addGestureRecognizer(tapGesture)
@@ -73,6 +74,9 @@ class ALKFriendVideoCell: ALKVideoCell {
 
         photoView.trailingAnchor.constraint(lessThanOrEqualTo: contentView.trailingAnchor, constant: -56).isActive = true
         photoView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -16).isActive = true
+
+        photoView.topAnchor.constraint(equalTo: nameLabel.bottomAnchor, constant: 6).isActive = true
+        photoView.widthAnchor.constraint(equalToConstant: width * 0.60).isActive = true
 
         timeLabel.leadingAnchor.constraint(equalTo: bubbleView.trailingAnchor, constant: 2).isActive = true
         timeLabel.bottomAnchor.constraint(equalTo: bubbleView.bottomAnchor, constant: 2).isActive = true